refactor(calendar): add typings for `trainingToDelete` useState
add styles for modal, change `Text` to `H2`, change labels

+++ src/screen/calendar/calendar.tsx
@@ -1,5 +1,5 @@
 import React, { useMemo, useState } from 'react';
-import { Button, StyleSheet, Text, View } from 'react-native';
+import { Button, StyleSheet, View } from 'react-native';
 import { connect, MapDispatchToPropsParam, MapStateToPropsParam } from 'react-redux';
 import { DEFAULT_DATE_FORMAT, formatDate, getToday } from '../../util/date.util';
 import moment from 'moment';
@@ -19,6 +19,8 @@ import { StoreModel } from '../../redux/store';
 import { deleteTrainingByIdAction } from '../../redux/action/training.action';
 import { useTranslation } from 'react-i18next';
 import Modal from 'react-native-modal';
+import { Colors } from '@css/colors.style';
+import { H2 } from '@components/heading/h2';
 
 interface IDispatchToProps {
 	deleteTrainingById: (trainingId: PropType<TrainingModel, 'id'>) => void;
@@ -36,32 +38,26 @@ const Calendar = (props: IProps & IStateToProps & IDispatchToProps) => {
 	const { t } = useTranslation();
 
 	const [selectedDate, changeSelectedDate] = useState(getToday());
-	const [trainingToDelete, changeTrainingToDelete] = useState();
+	const [trainingToDelete, changeTrainingToDelete] = useState<TrainingModel | undefined>();
 	const trainingList = useMemo(() => fetchTrainingListByDate(formatDate(selectedDate, DEFAULT_DATE_FORMAT)), [
 		fetchTrainingListByDate,
 		selectedDate,
 	]);
-
-	const styles = StyleSheet.create({
-		h1: {
-			fontSize: 24,
-			textAlign: 'center',
-			color: 'white',
-			paddingBottom: 24,
-		},
-	});
 
 	const handleOpenDeleteTrainingConfirm = (training: TrainingModel) => {
 		changeTrainingToDelete(training);
 	};
 
 	const handleCloseDeleteTrainingConfirm = () => {
-		changeTrainingToDelete('');
+		changeTrainingToDelete(undefined);
 	};
 
 	const deleteTraining = () => {
-		handleDeleteTraining(trainingToDelete);
-		changeTrainingToDelete(undefined);
+		if (trainingToDelete) {
+			handleDeleteTraining(trainingToDelete);
+		}
+
+		handleCloseDeleteTrainingConfirm();
 	};
 
 	const handleChangeSelectedDate = (date: moment.Moment) => {
@@ -99,10 +95,16 @@ const Calendar = (props: IProps & IStateToProps & IDispatchToProps) => {
 
 			<View>
 				<Modal isVisible={!!trainingToDelete}>
-					<View style={{ flex: 1 }}>
-						<Text style={styles.h1}>{t('Are you sure want to delete |item| ?', { item: trainingToDelete?.name })}</Text>
-						<Button title={t('Delete')} onPress={deleteTraining} />
-						<Button title={t('Hide modal')} onPress={handleCloseDeleteTrainingConfirm} />
+					<View style={styles.modal}>
+						<H2
+							style={styles.h2}
+							text={t('Are you sure want to delete |item| ?', { item: trainingToDelete?.name || '' })}
+						/>
+
+						<View style={styles.buttonWrapper}>
+							<Button color={Colors.LightRed} title={t('Delete')} onPress={deleteTraining} />
+							<Button title={t('Hide modal')} onPress={handleCloseDeleteTrainingConfirm} />
+						</View>
 					</View>
 				</Modal>
 			</View>
@@ -111,6 +113,21 @@ const Calendar = (props: IProps & IStateToProps & IDispatchToProps) => {
 		</View>
 	);
 };
+
+const styles = StyleSheet.create({
+	modal: {
+		paddingVertical: 10,
+		paddingHorizontal: 20,
+		backgroundColor: Colors.White,
+	},
+	h2: {
+		paddingBottom: 24,
+	},
+	buttonWrapper: {
+		flexDirection: 'row',
+		justifyContent: 'space-between',
+	},
+});
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const mapStateToProps: MapStateToPropsParam<IStateToProps, IProps, StoreModel> = (state, ownProps) => {
